Modified paging and test cases updated
Old pageable implementation had problems with filtering as when the list was filtered it calculated pages based on the unfiltered list. Switched to a dynamic implementation that applies after the list has been filtered, uses pointers for the first and last game that should be on each page.  Test cases needed to be modified to fit new implementation. Fixes all the filtering and paging problems.