Added spring security/authentication
Added spring security/authentication with register/login. Modified models to store the user for each entry to differentiate between users. Modified controllers to get and modify UserGame entries that are related to the logged in user only. Modified navbar to have a profile/logout section if logged in. Added a profile page with account information and stats.