Updated Javascript, modified toast functionality
Removed unnecessary variables that were only being used once.
Modified toasts to stack instead of replacing eachother, as well as removed the unnecessary arguments that were repeatedly passed in when calling showToast. The styling and icons are now calculated inside the method instead of directly being passed in as arguments.